requirements.txt
tensorflow
streamlit
numpy
matplotlib
imageio
ipython
Pillow
